+++ selfdrive/car/hyundai/carstate.py
@@ -99,10 +99,11 @@ class CarState(CarStateBase):
       ret.cruiseState.enabled = cp.vl["TCS13"]["ACC_REQ"] == 1
       ret.cruiseState.standstill = False
     else:
-      ret.cruiseState.available = cp_cruise.vl["SCC11"]["MainMode_ACC"] == 1
+      scc_bus = "SCC12" if self.CP.flags & HyundaiFlags.CAN_CANFD.value else "SCC11"
+      ret.cruiseState.available = cp_cruise.vl[scc_bus]["MainMode_ACC"] == 1
       ret.cruiseState.enabled = cp_cruise.vl["SCC12"]["ACCMode"] != 0
-      ret.cruiseState.standstill = cp_cruise.vl["SCC11"]["SCCInfoDisplay"] == 4.
-      ret.cruiseState.speed = cp_cruise.vl["SCC11"]["VSetDis"] * speed_conv
+      ret.cruiseState.standstill = cp_cruise.vl[scc_bus]["SCCInfoDisplay"] == 4.
+      ret.cruiseState.speed = cp_cruise.vl[scc_bus]["VSetDis"] * speed_conv
 
     # TODO: Find brake pressure
     ret.brake = 0
@@ -134,7 +135,7 @@ class CarState(CarStateBase):
 
     ret.gearShifter = self.parse_gear_shifter(self.shifter_values.get(gear))
 
-    if not self.CP.openpilotLongitudinalControl:
+    if not self.CP.openpilotLongitudinalControl and not (self.CP.flags & HyundaiFlags.CAN_CANFD.value):
       aeb_src = "FCA11" if self.CP.flags & HyundaiFlags.USE_FCA.value else "SCC12"
       aeb_sig = "FCA_CmdAct" if self.CP.flags & HyundaiFlags.USE_FCA.value else "AEB_CmdAct"
       aeb_warning = cp_cruise.vl[aeb_src]["CF_VSM_Warn"] != 0
@@ -291,7 +292,6 @@ class CarState(CarStateBase):
     ]
     checks = [
       # address, frequency
-      ("MDPS12", 50),
       ("TCS13", 50),
       ("TCS15", 10),
       ("CLU11", 50),
@@ -303,8 +303,10 @@ class CarState(CarStateBase):
       ("WHL_SPD11", 50),
       ("SAS11", 100),
     ]
+    freq_mdps12 = 100 if CP.flags & HyundaiFlags.CAN_CANFD.value else 50
+    checks.append(("MDPS12", freq_mdps12))
 
-    if not CP.openpilotLongitudinalControl and CP.carFingerprint not in CAMERA_SCC_CAR:
+    if not CP.openpilotLongitudinalControl and CP.carFingerprint not in CAMERA_SCC_CAR and not (CP.flags & HyundaiFlags.CAN_CANFD.value):
       signals += [
         ("MainMode_ACC", "SCC11"),
         ("VSetDis", "SCC11"),
@@ -330,13 +332,25 @@ class CarState(CarStateBase):
           ("CF_VSM_Warn", "SCC12"),
           ("CF_VSM_DecCmdAct", "SCC12"),
         ]
+    elif CP.flags & HyundaiFlags.CAN_CANFD.value:
+      signals += [
+        ("MainMode_ACC", "SCC12"),
+        ("VSetDis", "SCC12"),
+        ("SCCInfoDisplay", "SCC12"),
+        ("ACC_ObjDist", "SCC12"),
+        ("ACCMode", "SCC12"),
+      ]
+      checks += [
+        ("SCC12", 50),
+      ]
 
     if CP.enableBsm:
       signals += [
         ("CF_Lca_IndLeft", "LCA11"),
         ("CF_Lca_IndRight", "LCA11"),
       ]
-      checks.append(("LCA11", 50))
+      freq_lca11 = 20 if CP.flags & HyundaiFlags.CAN_CANFD.value else 50
+      checks.append(("LCA11", freq_lca11))
 
     if CP.carFingerprint in (HYBRID_CAR | EV_CAR):
       if CP.carFingerprint in HYBRID_CAR:
@@ -366,7 +380,8 @@ class CarState(CarStateBase):
       signals.append(("CF_Lvr_Gear", "LVR12"))
       checks.append(("LVR12", 100))
 
-    return CANParser(DBC[CP.carFingerprint]["pt"], signals, checks, 0)
+    bus = 4 if CP.flags & HyundaiFlags.CAN_CANFD.value else 0
+    return CANParser(DBC[CP.carFingerprint]["pt"], signals, checks, bus)
 
   @staticmethod
   def get_cam_can_parser(CP):
@@ -376,21 +391,24 @@ class CarState(CarStateBase):
     signals = [
       # signal_name, signal_address
       ("CF_Lkas_LdwsActivemode", "LKAS11"),
-      ("CF_Lkas_LdwsSysState", "LKAS11"),
-      ("CF_Lkas_SysWarning", "LKAS11"),
       ("CF_Lkas_LdwsLHWarning", "LKAS11"),
       ("CF_Lkas_LdwsRHWarning", "LKAS11"),
-      ("CF_Lkas_HbaLamp", "LKAS11"),
-      ("CF_Lkas_FcwBasReq", "LKAS11"),
-      ("CF_Lkas_HbaSysState", "LKAS11"),
-      ("CF_Lkas_FcwOpt", "LKAS11"),
-      ("CF_Lkas_HbaOpt", "LKAS11"),
-      ("CF_Lkas_FcwSysState", "LKAS11"),
-      ("CF_Lkas_FcwCollisionWarning", "LKAS11"),
-      ("CF_Lkas_FusionState", "LKAS11"),
       ("CF_Lkas_FcwOpt_USM", "LKAS11"),
-      ("CF_Lkas_LdwsOpt_USM", "LKAS11"),
     ]
+    if CP.flags & HyundaiFlags.CAN_CANFD.value:
+      signals += [
+        ("CF_Lkas_LdwsSysState", "LKAS11"),
+        ("CF_Lkas_SysWarning", "LKAS11"),
+        ("CF_Lkas_HbaLamp", "LKAS11"),
+        ("CF_Lkas_FcwBasReq", "LKAS11"),
+        ("CF_Lkas_HbaSysState", "LKAS11"),
+        ("CF_Lkas_FcwOpt", "LKAS11"),
+        ("CF_Lkas_HbaOpt", "LKAS11"),
+        ("CF_Lkas_FcwSysState", "LKAS11"),
+        ("CF_Lkas_FcwCollisionWarning", "LKAS11"),
+        ("CF_Lkas_FusionState", "LKAS11"),
+        ("CF_Lkas_LdwsOpt_USM", "LKAS11"),
+      ]
     checks = [
       ("LKAS11", 100)
     ]
@@ -422,7 +440,8 @@ class CarState(CarStateBase):
           ("CF_VSM_DecCmdAct", "SCC12"),
         ]
 
-    return CANParser(DBC[CP.carFingerprint]["pt"], signals, checks, 2)
+    bus = 6 if CP.flags & HyundaiFlags.CAN_CANFD else 2
+    return CANParser(DBC[CP.carFingerprint]["pt"], signals, checks, bus)
 
   @staticmethod
   def get_can_parser_canfd(CP):